+++ gui/objects/Colorizer.h
@@ -936,14 +936,14 @@ private:
 };
 
 /// \todo possibly move elsewhere
-static uint64_t getHash(const Size value, const Size seed) {
+static uint64_t getHash(const uint64_t value, const Size seed) {
     // https://stackoverflow.com/questions/8317508/hash-function-for-a-string
     constexpr int A = 54059;
     constexpr int B = 76963;
     constexpr int FIRST = 37;
 
     uint64_t hash = FIRST + seed;
-    StaticArray<uint8_t, sizeof(Size)> data;
+    StaticArray<uint8_t, sizeof(uint64_t)> data;
     std::memcpy(&data[0], &value, data.size());
     for (uint i = 0; i < sizeof(uint64_t); ++i) {
         hash = (hash * A) ^ (data[i] * B);